+++ public/service-worker.js
@@ -1,4 +1,4 @@
-const version = '1.2.0';
+const version = '1.3.1';
 
 const contentToCache = [
   '/',